requirements.txt
selenium>=4.6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0
rich>=13.0.0
tqdm>=4.65.0
